app/src/main/java/com/gasparaiciukas/owntrainer/database/ReminderDao.kt
package com.gasparaiciukas.owntrainer.database

import androidx.room.Dao
import androidx.room.Insert
import androidx.room.Query
import androidx.room.Update
import kotlinx.coroutines.flow.Flow

@Dao
interface ReminderDao {
    @Query("SELECT * FROM reminder")
    fun getReminders(): Flow<List<Reminder>>

    @Insert
    suspend fun insertReminder(reminder: Reminder)

    @Update
    suspend fun updateReminder(reminder: Reminder)

    @Query("DELETE FROM reminder WHERE reminderId = :id")
    suspend fun deleteReminderById(id: Int)
}